refactor(api): decompose social inbox service
