+++ apps/server/api/src/collections/social-inbox/social-inbox.module.ts
@@ -1,5 +1,9 @@
 import { SocialInboxController } from '@api/collections/social-inbox/controllers/social-inbox.controller';
 import { SocialInboxService } from '@api/collections/social-inbox/services/social-inbox.service';
+import { SocialInboxActionService } from '@api/collections/social-inbox/services/social-inbox-action.service';
+import { SocialInboxIngestionService } from '@api/collections/social-inbox/services/social-inbox-ingestion.service';
+import { SocialInboxQueryService } from '@api/collections/social-inbox/services/social-inbox-query.service';
+import { SocialInboxRealtimeService } from '@api/collections/social-inbox/services/social-inbox-realtime.service';
 import { WorkflowsModule } from '@api/collections/workflows/workflows.module';
 import { QueuesModule } from '@api/queues/core/queues.module';
 import { InstagramModule } from '@api/services/integrations/instagram/instagram.module';
@@ -17,6 +21,12 @@ import { forwardRef, Module } from '@nestjs/common';
     forwardRef(() => WorkflowsModule),
     forwardRef(() => YoutubeModule),
   ],
-  providers: [SocialInboxService],
+  providers: [
+    SocialInboxActionService,
+    SocialInboxIngestionService,
+    SocialInboxQueryService,
+    SocialInboxRealtimeService,
+    SocialInboxService,
+  ],
 })
 export class SocialInboxModule {}
